requirements.txt
# ─────────────────────────────────────────────
#  GarganDeFi Bot – requirements.txt
# ─────────────────────────────────────────────

# ── Core async framework ──────────────────────
asyncio==3.4.3; python_version < "3.10"   # stdlib shim for older envs
aiohttp>=3.9.5
aiofiles>=23.2.1

# ── Discord ───────────────────────────────────
discord.py>=2.4.0

# ── Solana / Web3 ─────────────────────────────
solana>=0.34.3
solders>=0.21.0
anchorpy>=0.20.1          # Raydium on-chain structs (optional but useful)

# ── HTTP scraping ─────────────────────────────
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.1
playwright>=1.44.0         # Headless browser fallback for JS-heavy sites
fake-useragent>=1.5.1

# ── Database ──────────────────────────────────
aiosqlite>=0.20.0

# ── Config & secrets ─────────────────────────
python-dotenv>=1.0.1

# ── Math & analytics ─────────────────────────
numpy>=1.26.4

# ── Streamlit dashboard ───────────────────────
streamlit>=1.35.0
pandas>=2.2.2
plotly>=5.22.0
altair>=5.3.0

# ── Scheduling / utilities ───────────────────
APScheduler>=3.10.4
tenacity>=8.3.0            # Retry decorator with exponential back-off
structlog>=24.1.0          # Structured async-safe logging

# ── Dev / testing ─────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.7

# ── Quant Analyst ─────────────────────────────
yfinance>=0.2.36
pandas_ta>=0.3.14b0
vaderSentiment>=3.3.2
feedparser>=6.0.11

# ── Reporting ───────────────────────────────
matplotlib>=3.8.4
seaborn>=0.13.2

# ── DevOps / Monitoring ─────────────────────
psutil>=5.9.8

# ── Extensions & AI ───────────────────────────
pybind11>=2.12.0           # For compiling C++ core_math module
google-generativeai>=0.4.0 # For Gemini AI routines
